Refactoring virtual_machine.go file
