--- vsphere/virtual_machine_test.go
@@ -1,1 +1,0 @@
-package vsphere
